README: add phpize and examples about extension
Another benefit of this nice framework is to be able to generate
and then to compile any standalone PHP extension (module).

It is mostly thanks to the phpize.bat tool that has been made
available.

TBC: should we add ext_skel.php and its template into the
deliverable of this SDK (c:\php\SDK\TBD) ?